Issue: For some API Server objects, API Server returns the error: "This object is not visible by users: XXX"
Fix:
Pass X_API_ROLE to the API Server along with X-AUTH-TOKEN, in API Server. Based
on this header, user visible entites are showed or thrown error.
When Sending to job server, user roles is passed along with other appData.
Added tokenObjs along with user Role list per project in session object, this
object is used to get the userRoles and pass to API Server.

Change-Id: Ia41fc83c1cfe09b779ef2c7b9997710e51ebab20