Related-Bug: #1419840
After creating a dvportgroup, need to get id for that dvportgroup such
that it can be associated to ip-pool.
For that, look-up calls need to be issued on all dvportgroups in that
datacenter, which is taking time.

Instead, maintain a cache of dvportgroup names and their ids on first
invocation, such that the lookup calls need to be issued for new
dvportgroups.

(cherry picked from commit e58aa3046fc602cd6329956799c66bb978900780)

Conflicts:
	src/serverroot/orchestration/plugins/vcenter/vcenter.api.js

Change-Id: If8e225004d08cc6b0ff99fbb405fde568e821ff0