Related-Bug: #1419840
After creating a dvportgroup, need to get id for that dvportgroup such
that it can be associated to ip-pool.
For that, look-up calls need to be issued on all dvportgroups in that
datacenter, which is taking time.

Instead, maintain a cache of dvportgroup names and their ids on first
invocation, such that the lookup calls need to be issued for new
dvportgroups.

Change-Id: I1c90bef856b2347c819b28073514581d6ba1bc55