Closes-Bug: #1438096
Currently, on webserver start, we get the dvs-id for the switchName
specified in config file and cache it for further SOAP requests.

But, have seen cases that dvs-id for a given switchname gets changed
(like the switch is deleted and re-created) and if it happens , UI tries
to create portgroup on an older dvs-id which doesn't exist.

In this case, modified to re-read the dvs-id for the specified switchName in config
file and continue with creating portgroup with new dvs-id.

Change-Id: I5e38b2e7ffc946c8463d5acf4dced462d07ead22
(cherry picked from commit b8c3936ebdac960fbb2138b7e19f586390d17b23)