Closes-Bug: #1439411
Earlier, chartOptions was moved outside initScatterChart function,as few
functions (bindEventsOverlapTooltip,showAxisParams) are moved outside of
initScatterChart and chartOptions is used in those functions.
But in some events (like click),chartOptions was being referred and we
need to refer to chartOptions that were provided while invoking
initScatterChart function.So, a single copy of chartOptions won't help.
Moved back chartOptions variable to initScatterChart function and passon
it as an argument to other functions if requried.

Change-Id: Icd0baaf9eabfbe0d2e4701e9d7f14289b3559217