Closes-Bug: #1460600
For Instances/Interfaces histograms in Dashboard,
the maximum value (max no of interfaces/instances on a vRouter) is divided by 8 to
get 8 equivalent buckets that represents 8 bars
Issue is the last bucket range includes only the maxValue - 1. So, the
vRoutes with maxValue of instnaces/interfaces are not considered while constructing the
histogram.

Fixed the issue by setting the last bucket range to maxValue + 1 such
that maxValue is included in the last bucket.

Change-Id: I796f01578573cdac04f8f58745f335684e90085d