Breadcrumb view with dropdown.
Added a function in such a way the url is accepted with the corresponding
parameter and render the dropdown and added to the breadcrumb.
Format of adding the breadcrumb

The function also take care of reading the cookie value and set the value for
the dropdown.

Sample code to add project and domain The parameter can be modified according to
the required dataset.

If new dataset is added 2 Case has to be added in the ControllerBreadcrumbView.
1. The way we render to read the value for the drop down.
2. To get the data of current value from the fqname passed as parameter.

self = this,
    hashParams = viewConfig.hashParams,
    domainSelected = {},
    fqName = null;

cobdcb.renderEachBreadcrumbDropdown(fqName, ctwc.URL_ALL_DOMAINS,
       ctwl.DOMAINS_BREADCRUMB_DROPDOWN, "domain",
       [ctwl.PROJECTS_BREADCRUMB_DROPDOWN],
       cowc.COOKIE_DOMAIN, ctwm.NO_DOMAIN_FOUND,
       function (selectedValueData, domainBreadcrumbChanged){
       if(selectedValueData != null && selectedValueData != "") {
         domainSelected = selectedValueData;
         cobdcb.renderEachBreadcrumbDropdown(fqName,
             ctwc.URL_CONFIG_PROJECT, ctwl.PROJECTS_BREADCRUMB_DROPDOWN,
             "project", [],cowc.COOKIE_PROJECT, ctwm.NO_PROJECT_FOUND,
             function(selectedValueData,projectBreadcrumbChanged) {
                 cowu.renderView4Config(this.$el, null,
                   this.getLogicalRouter(fqname, selectedValueData.value));
             }
         );
       }
       }
);

Closes-Bug: #1477845

Please enter the commit message for your changes. Lines starting

Change-Id: I084eb4527e43fa610e0274be1d154625e9cb0179