Relate-Bug: #1476413
testInitFn needs to be executed before mock data creation.
mock data creation expects the view object to be present in the rootview viewmap.
with the current renderOnActive configuration in the viewConfig, the tab views gets rendered only after clicking.
infra will pass a promise to wait till the init function is done. test page owner needs to resolve the promise based on the requirement.

Change-Id: I4663f1bccb76ed066bdc60812a215a8801674ab1