Relate-Bug: #1476413 UT infra feature additions and enhancements
1. testInitFn needs to be executed before mock data creation. (support dynamic tabs)
mock data creation expects the view object to be present in the rootview viewmap.
with the current renderOnActive configuration in the viewConfig, the tab views gets rendered only after clicking.
2. infra will pass a promise to wait till the init function is done.
test page owner needs to resolve the promise based on the requirement.
3. seperate model and view test runners
4. first cut for form model view tests.(More changes to come)
5. Common grid view test - detail view on row should not depend on particular cgrid.
6. renamed co.test.unit to co.test.runner. require config updated to co-test-runner to reference the same.
test page implementations needs to access this new ref
7. first cut for basic and lib api runners.
8. QUnit async calls should be enabled after library update.

Change-Id: I9412065fc57ebd62c38c06febf40ee3f86bf3fac