Related-Bug: #1544947 Currently Contrail WebUI checks for admin only role for all projects assigned to a user at login time.
Instead of checking all projects at login time, only check the assigned project for admin only
role and disallow login if user doesn't have admin only role in the assigned project.

RBAC is not supported by Contrail WebUI, member only projects cannot be used with Contrail UI.

At time of login, catalog response may contain high number of endpoints. Discard this data.

Change-Id: I07263eeef23e028740f2b9edfa6e34ac28e49362

Related-Bug: #1532666
As recommendation is to have admin role only, so if the last project is of
member role, then do not allow the user to login.
If after login, some user is found to be member, then as API Server returns
permission denied, so user will not be able to do anything, so user has to
logout, and then the available admin role project will be set as cookie.

Change-Id: I33c7370c4918f4bf1238bdb8c81ac3ddf8d4790b