Related-Bug: #1533202
Changes to show the range only from min value to max value so
that the chart area is used efficiently.
With this,we see the issue of bubbles that are plotted on both sides of
axis limits getting cropped (left/right or top/bottom).
So, added twice the maxCircleRadius to height & width to handle bubbles
that appear on both sides and translating the bubbles by maxCircleRadius.

Update only the grid header text with record count
Added separate getAxisMinMax for bucketization

Change-Id: I20d5cb3f30f8ff47a95e0f0766868a8efdb581da