Avoid passing an empty 'data' object to restler.get calls.
Restler changes Content-Type to 'application/x-www-form-urlencoded' even
for GET requests, as long as data is not a string. That breaks some nova
API calls when Mitaka nova-api is used, as it now expects to get
"application/json" for all queries. Make sure we pass data object only
for POST and PUT calls.

Change-Id: Ieaf1622dec703404d9672681a9cd0d35199326d7
Closes-Bug: 1591393