Closes-Bug: #1682252 If SMLite and contrail-webui is deployed in the same system, then they use same redis instance. So redis is shared across two middleware processes. So any publish on same channel goes to both.
Now in SMLite we do not have code execution path for jobs required for
controller. But SMLite subscribes for "job complete" event, now when controller
process completes that task, SMLite also gets notified on generic "job complete"
event, it processes that and creates a new Job (creates entry in redis to hand
over to blpop), but as the execution path is not registered by SMLite, SMLite
actually does not execute the job, it is done by controller, but all become
delayed jobs, so all delayed jobs pile up and then as part of job execution, it
send request to API/Op Server. So in this scenario, SMLite keeps on creating
multiple delayed jobs.

So fix is:
1. Do not do registration of "job complete" event if no of registered jobs is 0.
2. If a job is not registered, then do not create a job, this is required to
check if any APP has asked for job without registering.

Change-Id: I5adcc0cc5079caa6908a81bf6462f0926cbf370f