Closes-Bug: #1690888 For docker envt, we need to get the contrail-status through docker exec command. Now for docker envt, openstack, vrouter does not run in container, so we need to issue docker contrail-status command alongh with older contrail-status command. Now as nova compute runs along with vrouter node, so we need to issue openstack-status command where vrouter is running in case the request host is a compute node.
So basically we need to issue contrail-status/openstack-status using direct
command (for vrouter/openstack) and along with docker commands (for controller,
analytics and analyticsdb).

Now user can change the container names, UI does not have any knowledge of
that, so added config options to have the container names, default is controller,
analytics and analyticsdb.

Change-Id: I640fd73f0aa64ff2c51d9438a3f488c6fad221f1