Closes-Bug: #1698385 Fixed the loading issue that sometimes "Monitor > Networking > Dashboard" is shown on quickly jumping to "Traffic Groups" on expansion of "Monitor > Netorking" Menu
Issue is that on clicking a particular feature menu, it checks if it has
any initialization routing (mentioned as <init> in menu.xml) and if it's
there it loads that file and then starts rendering the feature page.
But in the meanwhile, if user has moved to another feature, still it
renders the old feature page.

Added the check to load the feature page after initialization routine
completes only if user hasn't moved to another page in between.

Change-Id: I8b5ca3051bbbdb4476c08d09cda51ee159b2cf39