Performance changes in network monitoring topology
Currently the topology page is crashing when there are
more than 400 networks after the following optimization
we are able to display 3000 networks in 10 - 15 secs.

Following optimizations helped,
1)Used the async flag provided by joint js which renders the elements in batches
and sorts once atlast.
2)Removed the extra markup added by our wrapper, now the DOM elements
reduced to half.
3)Removed the reference elements like text tag etc; which are placed with
reference to parent element, so whenever there is change in position of
parent element all the reference elements need to position again which leads
to rendering cycles.

Only cavet with these change is tooltips are bound only after all the
elements are rendered, so when the elements are rendering user can't see
the tooltips.

Related-Bug:#1644321

Change-Id: I237d4f2211798131082b60582c874451d328d464