Closes-Bug: #1702727 Adding Message of the day in all pages. Added file_path for this motd file, default is /etc/contrail/contrail-webui-motd.
Message of the day text alignment changes.

Now text is center aligned and left some space in left and right
such that it will not overlap with the menu & back to top link

Text container is always at the bottom of the page like footer,
user to scroll to end to see the message.

Change-Id: I5e996d0c476ae03060c07bff57b9c059aaeb2b43
Related-Bug:#1702727