Closes-Bug:#1715077 UI is not loading the proxyURL page after sucessful authentication
When we access the proxyURL like
https://nodeg3:8143/proxy?proxyURL=http://10.204.217.43:8083 without
login, ui displays the login page where we provide the login
credentials which should redirect to the proxyURL after successful
authentication.Currently instead of displaying the proxyURL page it is
displaying the default landing page.

Fix is in authenticate ajax call when referer header of the request object
contains the 'proxy' api we are reloading the page which displays the
corresponding the proxryURL page.

Change-Id: I55b6725d3225a5e8135d46161462bd5bb79b8149