Provide access to introspect using UI APIs without login
A new proxy request API is added in Contrail WebServer.
/forward-proxy

The format of the request is:
https://<web_server_ip>:8143/forward-proxy?<all_introspect_arguments>&proxyURL=http://<introspect_ip>:<introspect_port>
-H "X-Auth-Token: <scoped_token>"

If the request URL contains forward-proxy and X-Auth-Token header is not null,
then authentication is bypassed, we will assume these requests to be authenticated.
Now authorization will be done by the validity of token as specified in X-Auth-Token header.

In Contrail UI, we store admin_user, admin_password and admin_tenant_name in
/etc/contrail/contrail-webui-userauth.js Using all these, we will get admin-token
which will be used to validate the token as provided by user.

The keystone API to validate this token is:
/tokens/0123456789 -H "X-Auth-Token: <admin_token>"

If token is valid, then the response will include the details of the token and we pass the
request to application handler.

config.proxy.allowed_introspect_list_access_by_token contains the list of
allowed introspect names to access via keystone token.

config.proxy contains the allowed port list to be accessed via proxy.

Closes-Bug: #1716506

Change-Id: I5278953c1ab868ec03263e38fa0efd80eb8c6a7b