Fixing the query error issue from Monitor dashboard page while displaying the logs.
As we moved to the new query where the queries are sanitzed before sending to the
backend the queries with special characters lik < > were converted to &lt; and &rt;
which was causing issues in query engine.
Providing a quick fix by getting rid of the query by converting Level <= 4 to
Level = 1 OR Level = 2 etc.

Change-Id: I970ead2a4144d64121322b37a839f93daa6ecd70
Closes-Bug: #1722459