single sign on and html markup support in message of the day changes
Provided config option (connectedApp) in config.global.js on
enabling we should see links in footer of contrail webui which
takes to corresponding webpages using same tokens.

For other apps to contrail webui, we provided a api
/single-sign-on which accepts tokenid (scoped token) and
displays the contrail webui on successful authentication.

Added the html markup support in message of the day and moved
it to footer.
Related-Bug: #1727033, Closes-Bug: #1715469
Change-Id: Ibbf4c70d8292972c43211d2404510c7e73cffd78