Unable to login to webui if user and project belongs to different domains
Issue was while authenticating with keystone v3 in post data we used to use
domain as the user logged in, in stead of that we need to send the domain where
the project exists.

Now in stead of using project:name / domain:name in scope structure, using the
project:id, thus removing the need for domainname specify.

Change-Id: I5c4dcab8d9645dbabb5bdde54d99858d62e295c3
Closes-Bug: #1727809