Act nodeJS web-server as a request handler for GO Contrail Fabric UI
When request contains the x-auth-token, then skip the verify using csrf token.
verify that and get the details of the
token, and then populate tokenObjs, last_token_used and default_token to that.

So we should send API Server/Analytics request using the same token.

So with this change, we will have two ways to verify UI request, either CSRF (older
way) or X-Auth-Token.

Change-Id: Ibb6277e2a485b2cc63136a18a2e8752bf3bb0909
Depends-On: Iaf6b0bab9ca215d74d73c95f8ea0e35609faabac
Closes-Bug: #1745048