Added ability to add dependency on required packages for menu items and tabs
1. We can now specify the required packages for every menu item. The required
package names are assumed be environment variables in the container.
For eg. if Physical Topology is dependent on ANALYTICS_SNMP_NODES environment
variable we can make ensure to enable this menu item by specifying the
following in access section of the menu.xml

                            <access>
                                <reqdPackages>
                                    <package>ANALYTICS_SNMP_NODES</package>
                                </reqdPackages>
                            </access>

The MenuHandler will check for this environment variable by looking for in the
webServerInfo global variable and enable the menu only if present.

2. We can now also put a dependency on a package/env variable for the tabs as well.
For eg.
tabConfig: {
    reqdPackages: cowc.ANALYTICS_QUERY_PACKAGES
}

3. For nodejs APIs we can specify a dependency in the parseURL.xml and make an API return
a predefined response or return null without making a call to the OpServer or other
backend.
For eg.
        <reqdPackages>
            <packages>
                <package>ANALYTICSDB_NODES</package>
            </packages>
            <noPackageResponse>{"data":[]}</noPackageResponse>
        </reqdPackages>

Also changes to display the bell icon only when the required analytics component is installed.

Change-Id: Iee2fe1b7a7f789df033d06533e64cc559b4f10d3
Related-Bug: #1792645