Fix Title: AIC_L RC26.2 IBM AppScan - DYH3a Session Identifier Not Updated - Contrail 3.2.16.0-117
Problem statement:
This  is issue found in the IBM App Scan in AT&T.
It may be possible to steal or manipulate customer session and cookies, which might be used to impersonate a legitimate user, allowing the hacker to view or alter user records, and to perform transactions as that user.

Proposed fix:
Upon each successful login/authentication, we are regenerating the session and populating necessary data on this new session object
and upon each logout, we are destroying the session object in store and clearing the session id in cookie.

Testing done: yes

Closes-Jira-Bug: CEM-21250
Signed-off-by: root <ralqua@juniper.net>
Change-Id: I4b47cd40b16e9d07f517000d9b186d2820f988ee