Fix Title: Tungsten UI session is lost after upgrade to 1912.L4
Problem statement:
After successful login to Tungsten UI, after a second or two user is kicked out to login screen.
Seems that session is not properly handled between nodes.

Proposed fix:
Upon each successful login/auth, we are regenerating the session and populating necessary data on this new session object
and upon each logout, we are destroying the session object in store and clearing the session id in cookie

Note: this a fix attempt based on reasoning/guess. Root cause analysis
faces challenges due to the fact that restarting the webUI containers will
eliminated the issue.

partial-jira-bug: CEM-22659
Signed-off-by: Rabbani <ralqua@juniper.net>
Change-Id: Ia21a528e244876f129704ac66dc4934bbc092aad